--- server.bat
@@ -1,3 +1,0 @@
-@echo off
-python server.py
-pause
